perf: shrink OpCode from u16 to u8 in change buffer protocol
OpCode values currently range 0–16, well within u8 (0–255).
Shrinking the on-wire opcode from 2 bytes to 1 saves 1 byte per
operation in the change buffer — a ~11% reduction in header overhead
per op (opcode byte + span_id u64 = 9B vs 10B).

+++ libdd-trace-utils/src/change_buffer/mod.rs
@@ -7,7 +7,7 @@ pub enum ChangeBufferError {
     StringNotFound(u16),
     ReadOutOfBounds { offset: usize, len: usize },
     WriteOutOfBounds { offset: usize, len: usize },
-    UnknownOpcode(u16),
+    UnknownOpcode(u8),
 }
 
 impl std::fmt::Display for ChangeBufferError {
